CLI Import - prepare_import_data() refactor
* move prepare_import_data() out of helpers (very specific to be there)
* move all dataset preparations into prepare_import_data()
* fix misnomer for custom-channels
